requirements.txt
streamlit==1.31.0
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.17.2
groq==0.4.2
pypdf2==3.0.1
python-dotenv==1.0.0
webdriver-manager==4.0.1